requirements.txt
langchain==0.0.154
langchain_groq
gorq
PyPDF2==3.0.1
python-dotenv==1.0.0
fastembed
chromadb
streamlit==1.18.1
streamlit-extras